separate latest_validator_vote_packet.rs
